workflow-manager/pyproject.toml
[project]
name = "workflow_manager"
version = "0.1.0"
description = "Workflow task management system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.13.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "psycopg2-binary>=2.9.10",
    "grpcio>=1.67.0",
    "grpcio-tools>=1.67.0",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.2",
    "croniter>=3.0.3",
    "apscheduler>=3.10.4",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/workflow_manager"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",  # Line too long
    "N802",  # Function name should be lowercase (gRPC methods use PascalCase)
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
